webapi/webapi-webrtc-w3c-tests/webrtc/blink/RTCPeerConnection-AddRemoveStream.html

<!DOCTYPE html>
<!--
Test converted from blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/mediastream/RTCPeerConnection-AddRemoveStream.html
-->

<meta charset=utf-8>
<title>WebRTC Test: RTCPeerConnection [add|remove]Stream</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel=help href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcpeerconnection-interface">
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<div id=log></div>
<script>

var stream = null;
var stream2 = null;
var pc = null;

var t = async_test(document.title, {timeout: 5000});

function error() {
  t.step(function () {
    assert_unreached("Stream generation failed.");
  });
  t.done();
}

function getUserMedia(dictionary, callback) {
  t.step(function () {
    if (typeof (navigator.getUserMedia) != "undefined") {
      navigator.getUserMedia(dictionary, callback, error);
    } else {
      navigator.webkitGetUserMedia(dictionary, callback, error);
    }
  });
}

function onErroneousNegotiationNeeded() {
  t.step(function () {
    assert_unreached("onErroneousNegotiationNeeded was called.");
  });
  t.done();
}

function onRemoveStream(event) {
  t.step(function () {
    assert_equals(pc.getLocalStreams().length, 0);
  });
  t.done();
}

function onAddStream(event) {
  t.step(function () {
    assert_equals(pc.getStreamById(stream.id), stream);

    assert_equals(pc.getStreamById(stream2.id), null);

    pc.onnegotiationneeded = onErroneousNegotiationNeeded;
    pc.addStream(stream);

    assert_equals(pc.getLocalStreams().length, 1);

    pc.removeStream(stream2);

    assert_equals(pc.getLocalStreams().length, 1);
  });

  pc.onnegotiationneeded = onRemoveStream;
  pc.removeStream(stream);
}

function gotStream2(s) {
  stream2 = s;
  t.step(function () {
    assert_not_equals(stream.id, stream2.id);
    if (typeof RTCPeerConnection != "undefined") {
      pc = new RTCPeerConnection(null, null);
    } else {
      pc = new webkitRTCPeerConnection(null, null);
    }
    pc.onnegotiationneeded = onAddStream;
    pc.addStream(stream);
  });
}

function gotStream1(s) {
  t.step(function () {
    stream = s;
    getUserMedia({audio:true, video:true}, gotStream2);
  });
}

getUserMedia({audio:true, video:true}, gotStream1);

</script>
